Convert database skills enum to table
Converted the skills enum in Postgres to a table of all possible skills,
updating the skills table to use a foreign key constraint on the skill
column, instead of a native enum or check statement.

This will make modifying the set of skills a lot simpler - skills can
simply be inserted/deleted/updated with INSERT/DELETE/UPDATE SQL
statements on the skill_enum table.

Also updated knex, pg, and modified docker-compose.yml to include
default settings for environment variables.